fix: test setup using snapshots
