--- packages/eds-tokens-build/src/__tests__/expected/color-scheme.css
@@ -1,110 +1,0 @@
-/**
- * Do not edit directly, this file was auto-generated by scripts/build-color-scheme-variables.ts.
- */
-
-:root,
-[data-color-scheme="light"] {
-  color-scheme: light;
-}
-
-[data-color-scheme="dark"] {
-  color-scheme: dark;
-}
-
-:root {
-  --eds-color-moss-green-1: light-dark(#eaf8fa, #0a0b0b);
-  --eds-color-moss-green-2: light-dark(#f6ffff, #1e2323);
-  --eds-color-moss-green-3: light-dark(#c5e1e4, #3c6266);
-  --eds-color-moss-green-4: light-dark(#b1d6da, #3e7378);
-  --eds-color-moss-green-5: light-dark(#97c8cd, #41878e);
-  --eds-color-moss-green-6: light-dark(#bbdbdf, #3c6266);
-  --eds-color-moss-green-7: light-dark(#7cbac1, #439199);
-  --eds-color-moss-green-8: light-dark(#21767e, #6ec0c9);
-  --eds-color-moss-green-9: light-dark(#206f77, #8cd2da);
-  --eds-color-moss-green-10: light-dark(#205c62, #ace3e9);
-  --eds-color-moss-green-11: light-dark(#20565c, #c7f1f6);
-  --eds-color-moss-green-12: light-dark(#1f6369, #bcebf1);
-  --eds-color-moss-green-13: light-dark(#141f20, #e6ffff);
-  --eds-color-moss-green-14: light-dark(#cae4e7, #2c3839);
-  --eds-color-moss-green-15: light-dark(#ffffff, #030303);
-  --eds-color-gray-1: light-dark(#f5f5f5, #0b0b0b);
-  --eds-color-gray-2: light-dark(#ffffff, #222222);
-  --eds-color-gray-3: light-dark(#dbdbdb, #5b5b5b);
-  --eds-color-gray-4: light-dark(#cecece, #696969);
-  --eds-color-gray-5: light-dark(#bebebe, #7a7a7a);
-  --eds-color-gray-6: light-dark(#d4d4d4, #5b5b5b);
-  --eds-color-gray-7: light-dark(#aeaeae, #838383);
-  --eds-color-gray-8: light-dark(#696969, #b1b1b1);
-  --eds-color-gray-9: light-dark(#636363, #c4c4c4);
-  --eds-color-gray-10: light-dark(#525252, #d7d7d7);
-  --eds-color-gray-11: light-dark(#4d4d4d, #e8e8e8);
-  --eds-color-gray-12: light-dark(#585858, #e1e1e1);
-  --eds-color-gray-13: light-dark(#1d1d1d, #fcfcfc);
-  --eds-color-gray-14: light-dark(#dedede, #353535);
-  --eds-color-gray-15: light-dark(#ffffff, #030303);
-  --eds-color-green-1: light-dark(#eafbe8, #0a0c0a);
-  --eds-color-green-2: light-dark(#f6fff5, #1e231e);
-  --eds-color-green-3: light-dark(#c5e5c2, #3c673a);
-  --eds-color-green-4: light-dark(#b1dbae, #3e793c);
-  --eds-color-green-5: light-dark(#97cf93, #418e3e);
-  --eds-color-green-6: light-dark(#bbe0b8, #3c673a);
-  --eds-color-green-7: light-dark(#7cc278, #439941);
-  --eds-color-green-8: light-dark(#227e22, #6eca6a);
-  --eds-color-green-9: light-dark(#207720, #8cdb87);
-  --eds-color-green-10: light-dark(#20621f, #aceba8);
-  --eds-color-green-11: light-dark(#215c1f, #c7f7c3);
-  --eds-color-green-12: light-dark(#20691f, #bcf2b8);
-  --eds-color-green-13: light-dark(#142114, #e6ffe3);
-  --eds-color-green-14: light-dark(#cae8c7, #2c392b);
-  --eds-color-green-15: light-dark(#ffffff, #030303);
-  --eds-color-blue-1: light-dark(#e7f8ff, #0a0b0c);
-  --eds-color-blue-2: light-dark(#f4ffff, #1d2226);
-  --eds-color-blue-3: light-dark(#c0e0f8, #33607e);
-  --eds-color-blue-4: light-dark(#aad4f3, #316f98);
-  --eds-color-blue-5: light-dark(#8dc5ee, #2e82b7);
-  --eds-color-blue-6: light-dark(#b5daf5, #33607e);
-  --eds-color-blue-7: light-dark(#6fb6e9, #2d8bc5);
-  --eds-color-blue-8: light-dark(#0070a9, #5abbfb);
-  --eds-color-blue-9: light-dark(#006aa0, #7dceff);
-  --eds-color-blue-10: light-dark(#085883, #a2e0ff);
-  --eds-color-blue-11: light-dark(#0e5279, #c4eeff);
-  --eds-color-blue-12: light-dark(#015e8d, #b7e8ff);
-  --eds-color-blue-13: light-dark(#121e27, #ecffff);
-  --eds-color-blue-14: light-dark(#c5e3f9, #2a3741);
-  --eds-color-blue-15: light-dark(#ffffff, #030304);
-  --eds-color-orange-1: light-dark(#fff1e2, #0c0b0a);
-  --eds-color-orange-2: light-dark(#fffcf0, #27201b);
-  --eds-color-orange-3: light-dark(#f9d2b6, #7e4e25);
-  --eds-color-orange-4: light-dark(#f4c29d, #97571b);
-  --eds-color-orange-5: light-dark(#eeae7c, #b46201);
-  --eds-color-orange-6: light-dark(#f6caaa, #7e4e25);
-  --eds-color-orange-7: light-dark(#e89959, #c36800);
-  --eds-color-orange-8: light-dark(#a34e00, #f99539);
-  --eds-color-orange-9: light-dark(#9b4900, #ffad63);
-  --eds-color-orange-10: light-dark(#813e00, #ffc791);
-  --eds-color-orange-11: light-dark(#773a00, #ffddb9);
-  --eds-color-orange-12: light-dark(#8a4100, #ffd4aa);
-  --eds-color-orange-13: light-dark(#27190e, #fff7e6);
-  --eds-color-orange-14: light-dark(#fad6bc, #413226);
-  --eds-color-orange-15: light-dark(#ffffff, #040303);
-  --eds-color-red-1: light-dark(#ffecea, #0d0a0a);
-  --eds-color-red-2: light-dark(#fff9f8, #2a1e1e);
-  --eds-color-red-3: light-dark(#ffc6c4, #923a3c);
-  --eds-color-red-4: light-dark(#ffb2b0, #b03940);
-  --eds-color-red-5: light-dark(#ff9897, #d43745);
-  --eds-color-red-6: light-dark(#ffbcba, #923a3c);
-  --eds-color-red-7: light-dark(#ff7a7d, #e53748);
-  --eds-color-red-8: light-dark(#c6002d, #ff8082);
-  --eds-color-red-9: light-dark(#bc002a, #ffa3a1);
-  --eds-color-red-10: light-dark(#9a1026, #ffc1bf);
-  --eds-color-red-11: light-dark(#8e1525, #ffd9d7);
-  --eds-color-red-12: light-dark(#a50827, #ffd0ce);
-  --eds-color-red-13: light-dark(#2e1414, #fff5f4);
-  --eds-color-red-14: light-dark(#ffcbc9, #492d2c);
-  --eds-color-red-15: light-dark(#ffffff, #040303);
-  --eds-color-bg-floating: light-dark(#ffffff, #5b5b5b);
-  --eds-color-bg-backdrop: light-dark(#aeaeae, #222222);
-  --eds-color-bg-input: light-dark(#f5f5f5, #0b0b0b);
-  --eds-color-border-focus: light-dark(#6fb6e9, #2d8bc5);
-  --eds-color-text-link: light-dark(#121e27, #ecffff);
-}
