--- packages/eds-tokens-build/src/__tests__/expected/dynamic-variables.css
@@ -1,138 +1,0 @@
-/**
- * Do not edit directly, this file was auto-generated by scripts/build-semantic-dynamic-variables.ts.
- */
-
-
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-:root, [data-color-appearance="neutral"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-gray-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-gray-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-gray-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-gray-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-gray-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-gray-11);
-  --eds-color-bg-canvas: var(--eds-color-gray-1);
-  --eds-color-bg-surface: var(--eds-color-gray-2);
-  --eds-color-border-subtle: var(--eds-color-gray-6);
-  --eds-color-border-medium: var(--eds-color-gray-7);
-  --eds-color-border-strong: var(--eds-color-gray-8);
-  --eds-color-text-subtle: var(--eds-color-gray-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-gray-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-gray-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-gray-15); /** Text or icons against colored backgrounds */
-}
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-[data-color-appearance="accent"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-moss-green-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-moss-green-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-moss-green-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-moss-green-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-moss-green-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-moss-green-11);
-  --eds-color-bg-canvas: var(--eds-color-moss-green-1);
-  --eds-color-bg-surface: var(--eds-color-moss-green-2);
-  --eds-color-border-subtle: var(--eds-color-moss-green-6);
-  --eds-color-border-medium: var(--eds-color-moss-green-7);
-  --eds-color-border-strong: var(--eds-color-moss-green-8);
-  --eds-color-text-subtle: var(--eds-color-moss-green-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-moss-green-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-moss-green-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-moss-green-15); /** Text or icons against colored backgrounds */
-}
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-[data-color-appearance="danger"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-red-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-red-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-red-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-red-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-red-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-red-11);
-  --eds-color-bg-canvas: var(--eds-color-red-1);
-  --eds-color-bg-surface: var(--eds-color-red-2);
-  --eds-color-border-subtle: var(--eds-color-red-6);
-  --eds-color-border-medium: var(--eds-color-red-7);
-  --eds-color-border-strong: var(--eds-color-red-8);
-  --eds-color-text-subtle: var(--eds-color-red-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-red-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-red-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-red-15); /** Text or icons against colored backgrounds */
-}
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-[data-color-appearance="info"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-blue-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-blue-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-blue-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-blue-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-blue-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-blue-11);
-  --eds-color-bg-canvas: var(--eds-color-blue-1);
-  --eds-color-bg-surface: var(--eds-color-blue-2);
-  --eds-color-border-subtle: var(--eds-color-blue-6);
-  --eds-color-border-medium: var(--eds-color-blue-7);
-  --eds-color-border-strong: var(--eds-color-blue-8);
-  --eds-color-text-subtle: var(--eds-color-blue-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-blue-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-blue-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-blue-15); /** Text or icons against colored backgrounds */
-}
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-[data-color-appearance="success"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-green-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-green-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-green-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-green-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-green-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-green-11);
-  --eds-color-bg-canvas: var(--eds-color-green-1);
-  --eds-color-bg-surface: var(--eds-color-green-2);
-  --eds-color-border-subtle: var(--eds-color-green-6);
-  --eds-color-border-medium: var(--eds-color-green-7);
-  --eds-color-border-strong: var(--eds-color-green-8);
-  --eds-color-text-subtle: var(--eds-color-green-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-green-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-green-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-green-15); /** Text or icons against colored backgrounds */
-}
-
-/**
- * Do not edit directly, this file was auto-generated.
- */
-
-[data-color-appearance="warning"] {
-  --eds-color-bg-fill-muted-default: var(--eds-color-orange-3);
-  --eds-color-bg-fill-muted-hover: var(--eds-color-orange-4);
-  --eds-color-bg-fill-muted-active: var(--eds-color-orange-5);
-  --eds-color-bg-fill-emphasis-default: var(--eds-color-orange-9);
-  --eds-color-bg-fill-emphasis-hover: var(--eds-color-orange-10);
-  --eds-color-bg-fill-emphasis-active: var(--eds-color-orange-11);
-  --eds-color-bg-canvas: var(--eds-color-orange-1);
-  --eds-color-bg-surface: var(--eds-color-orange-2);
-  --eds-color-border-subtle: var(--eds-color-orange-6);
-  --eds-color-border-medium: var(--eds-color-orange-7);
-  --eds-color-border-strong: var(--eds-color-orange-8);
-  --eds-color-text-subtle: var(--eds-color-orange-12); /** Used for text and icons */
-  --eds-color-text-strong: var(--eds-color-orange-13); /** Used for text and icons */
-  --eds-color-text-subtle-on-emphasis: var(--eds-color-orange-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-strong-on-emphasis: var(--eds-color-orange-15); /** Text or icons against colored backgrounds */
-}
-
