--- packages/eds-tokens-build/src/__tests__/expected/static-variables.css
@@ -1,97 +1,0 @@
-/**
- * Do not edit directly, this file was auto-generated by scripts/build-static-variables.ts.
- * Note: color scheme variables have moved to color-scheme.css. Import both files.
- */
-
-:root {
-  --eds-color-bg-neutral-canvas: var(--eds-color-gray-1);
-  --eds-color-bg-neutral-surface: var(--eds-color-gray-2);
-  --eds-color-bg-neutral-fill-muted-default: var(--eds-color-gray-3);
-  --eds-color-bg-neutral-fill-muted-hover: var(--eds-color-gray-4);
-  --eds-color-bg-neutral-fill-muted-active: var(--eds-color-gray-5);
-  --eds-color-bg-neutral-fill-emphasis-default: var(--eds-color-gray-9);
-  --eds-color-bg-neutral-fill-emphasis-hover: var(--eds-color-gray-10);
-  --eds-color-bg-neutral-fill-emphasis-active: var(--eds-color-gray-11);
-  --eds-color-bg-accent-canvas: var(--eds-color-moss-green-1);
-  --eds-color-bg-accent-surface: var(--eds-color-moss-green-2);
-  --eds-color-bg-accent-fill-muted-default: var(--eds-color-moss-green-3);
-  --eds-color-bg-accent-fill-muted-hover: var(--eds-color-moss-green-4);
-  --eds-color-bg-accent-fill-muted-active: var(--eds-color-moss-green-5);
-  --eds-color-bg-accent-fill-emphasis-default: var(--eds-color-moss-green-9);
-  --eds-color-bg-accent-fill-emphasis-hover: var(--eds-color-moss-green-10);
-  --eds-color-bg-accent-fill-emphasis-active: var(--eds-color-moss-green-11);
-  --eds-color-bg-success-canvas: var(--eds-color-green-1);
-  --eds-color-bg-success-surface: var(--eds-color-green-2);
-  --eds-color-bg-success-fill-muted-default: var(--eds-color-green-3);
-  --eds-color-bg-success-fill-muted-hover: var(--eds-color-green-4);
-  --eds-color-bg-success-fill-muted-active: var(--eds-color-green-5);
-  --eds-color-bg-success-fill-emphasis-default: var(--eds-color-green-9);
-  --eds-color-bg-success-fill-emphasis-hover: var(--eds-color-green-10);
-  --eds-color-bg-success-fill-emphasis-active: var(--eds-color-green-11);
-  --eds-color-bg-info-canvas: var(--eds-color-blue-1);
-  --eds-color-bg-info-surface: var(--eds-color-blue-2);
-  --eds-color-bg-info-fill-muted-default: var(--eds-color-blue-3);
-  --eds-color-bg-info-fill-muted-hover: var(--eds-color-blue-4);
-  --eds-color-bg-info-fill-muted-active: var(--eds-color-blue-5);
-  --eds-color-bg-info-fill-emphasis-default: var(--eds-color-blue-9);
-  --eds-color-bg-info-fill-emphasis-hover: var(--eds-color-blue-10);
-  --eds-color-bg-info-fill-emphasis-active: var(--eds-color-blue-11);
-  --eds-color-bg-warning-canvas: var(--eds-color-orange-1);
-  --eds-color-bg-warning-surface: var(--eds-color-orange-2);
-  --eds-color-bg-warning-fill-muted-default: var(--eds-color-orange-3);
-  --eds-color-bg-warning-fill-muted-hover: var(--eds-color-orange-4);
-  --eds-color-bg-warning-fill-muted-active: var(--eds-color-orange-5);
-  --eds-color-bg-warning-fill-emphasis-default: var(--eds-color-orange-9);
-  --eds-color-bg-warning-fill-emphasis-hover: var(--eds-color-orange-10);
-  --eds-color-bg-warning-fill-emphasis-active: var(--eds-color-orange-11);
-  --eds-color-bg-danger-canvas: var(--eds-color-red-1);
-  --eds-color-bg-danger-surface: var(--eds-color-red-2);
-  --eds-color-bg-danger-fill-muted-default: var(--eds-color-red-3);
-  --eds-color-bg-danger-fill-muted-hover: var(--eds-color-red-4);
-  --eds-color-bg-danger-fill-muted-active: var(--eds-color-red-5);
-  --eds-color-bg-danger-fill-emphasis-default: var(--eds-color-red-9);
-  --eds-color-bg-danger-fill-emphasis-hover: var(--eds-color-red-10);
-  --eds-color-bg-danger-fill-emphasis-active: var(--eds-color-red-11);
-  --eds-color-border-neutral-subtle: var(--eds-color-gray-6);
-  --eds-color-border-neutral-medium: var(--eds-color-gray-7);
-  --eds-color-border-neutral-strong: var(--eds-color-gray-8);
-  --eds-color-border-accent-subtle: var(--eds-color-moss-green-6);
-  --eds-color-border-accent-medium: var(--eds-color-moss-green-7);
-  --eds-color-border-accent-strong: var(--eds-color-moss-green-8);
-  --eds-color-border-success-subtle: var(--eds-color-green-6);
-  --eds-color-border-success-medium: var(--eds-color-green-7);
-  --eds-color-border-success-strong: var(--eds-color-green-8);
-  --eds-color-border-info-subtle: var(--eds-color-blue-6);
-  --eds-color-border-info-medium: var(--eds-color-blue-7);
-  --eds-color-border-info-strong: var(--eds-color-blue-8);
-  --eds-color-border-warning-subtle: var(--eds-color-orange-6);
-  --eds-color-border-warning-medium: var(--eds-color-orange-7);
-  --eds-color-border-warning-strong: var(--eds-color-orange-8);
-  --eds-color-border-danger-subtle: var(--eds-color-red-6);
-  --eds-color-border-danger-medium: var(--eds-color-red-7);
-  --eds-color-border-danger-strong: var(--eds-color-red-8);
-  --eds-color-text-neutral-subtle: var(--eds-color-gray-12); /** Used for text and icons */
-  --eds-color-text-neutral-strong: var(--eds-color-gray-13); /** Used for text and icons */
-  --eds-color-text-neutral-subtle-on-emphasis: var(--eds-color-gray-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-neutral-strong-on-emphasis: var(--eds-color-gray-15); /** Text or icons against colored backgrounds */
-  --eds-color-text-accent-subtle: var(--eds-color-moss-green-12); /** Used for text and icons */
-  --eds-color-text-accent-strong: var(--eds-color-moss-green-13); /** Used for text and icons */
-  --eds-color-text-accent-subtle-on-emphasis: var(--eds-color-moss-green-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-accent-strong-on-emphasis: var(--eds-color-moss-green-15); /** Text or icons against colored backgrounds */
-  --eds-color-text-success-subtle: var(--eds-color-green-12); /** Used for text and icons */
-  --eds-color-text-success-strong: var(--eds-color-green-13); /** Used for text and icons */
-  --eds-color-text-success-subtle-on-emphasis: var(--eds-color-green-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-success-strong-on-emphasis: var(--eds-color-green-15); /** Text or icons against colored backgrounds */
-  --eds-color-text-info-subtle: var(--eds-color-blue-12); /** Used for text and icons */
-  --eds-color-text-info-strong: var(--eds-color-blue-13); /** Used for text and icons */
-  --eds-color-text-info-subtle-on-emphasis: var(--eds-color-blue-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-info-strong-on-emphasis: var(--eds-color-blue-15); /** Text or icons against colored backgrounds */
-  --eds-color-text-warning-subtle: var(--eds-color-orange-12); /** Used for text and icons */
-  --eds-color-text-warning-strong: var(--eds-color-orange-13); /** Used for text and icons */
-  --eds-color-text-warning-subtle-on-emphasis: var(--eds-color-orange-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-warning-strong-on-emphasis: var(--eds-color-orange-15); /** Text or icons against colored backgrounds */
-  --eds-color-text-danger-subtle: var(--eds-color-red-12); /** Used for text and icons */
-  --eds-color-text-danger-strong: var(--eds-color-red-13); /** Used for text and icons */
-  --eds-color-text-danger-subtle-on-emphasis: var(--eds-color-red-14); /** Text or icons against colored backgrounds */
-  --eds-color-text-danger-strong-on-emphasis: var(--eds-color-red-15); /** Text or icons against colored backgrounds */
-}
