--- packages/eds-tokens-build/src/__tests__/fixtures/test-token-config.json
@@ -1,22 +1,0 @@
-{
-  "figmaProjectFoundationId": "test-foundation",
-  "figmaProjectStaticId": "test-static",
-  "figmaProjectDynamicId": "test-dynamic",
-  "schemeTokensPrefix": "🌗 Color Scheme",
-  "variablePrefix": "test",
-  "buildPath": "color/",
-  "semanticColorCategories": {
-    "Accent": "Brand",
-    "Neutral": "Gray"
-  },
-  "conceptColorGroups": {
-    "bg-floating": {
-      "Light": "{Light.Gray.1}",
-      "Dark": "{Dark.Gray.1}"
-    },
-    "border-focus": {
-      "Light": "{Light.Blue.7}",
-      "Dark": "{Dark.Blue.7}"
-    }
-  }
-}
